+++ source/Metadata/Class.ts
@@ -10,7 +10,7 @@ import { ECObjectsError, ECObjectsStatus } from "../Exception";
 import { PrimitiveProperty, PrimitiveArrayProperty, StructProperty, StructArrayProperty, EnumerationProperty, EnumerationArrayProperty, Property } from "./Property";
 import { DelayedPromiseWithProps } from "../DelayedPromise";
 import Schema from "./Schema";
-import { AnyClass, LazyLoadedECClass, SchemaItemVisitor } from "../Interfaces";
+import { AnyClass, LazyLoadedECClass, SchemaItemVisitor, LazyLoadedSchemaItem } from "../Interfaces";
 
 function createLazyLoadedItem<T extends SchemaItem>(c: T) {
   return new DelayedPromiseWithProps(c.key, async () => c);
@@ -57,20 +57,24 @@ export default abstract class ECClass extends SchemaItem implements CustomAttrib
     return prop;
   }
 
-  public getBaseClassSync(): ECClass | undefined {
-    if (!this.baseClass)
+  protected getReferencedClassSync<T extends ECClass>(key?: LazyLoadedSchemaItem<T>): T | undefined {
+    if (!key)
       return undefined;
 
-    const isInThisSchema = (this.schema.name.toLowerCase() === this.baseClass.schemaName.toLowerCase());
+    const isInThisSchema = (this.schema.name.toLowerCase() === key.schemaName.toLowerCase());
 
     if (isInThisSchema)
-      return this.schema.getClassSync(this.baseClass.name);
+      return this.schema.getClassSync<T>(key.name);
 
-    const reference = this.schema.getReferenceSync(this.baseClass.schemaName);
+    const reference = this.schema.getReferenceSync(key.schemaName);
     if (reference)
-      return reference.getClassSync(this.baseClass.name);
+      return reference.getClassSync<T>(key.name);
 
     return undefined;
+  }
+
+  public getBaseClassSync(): ECClass | undefined {
+    return this.getReferencedClassSync(this.baseClass);
   }
 
   /**
@@ -423,6 +427,28 @@ export default abstract class ECClass extends SchemaItem implements CustomAttrib
       await addBaseClasses(baseClass);
       if (baseClass !== this)
         yield baseClass;
+    }
+  }
+
+public *getAllBaseClassesSync(): Iterable<AnyClass> {
+  const baseClasses: ECClass[] = [ this ];
+  const addBaseClasses = (ecClass: AnyClass) => {
+    if (SchemaItemType.EntityClass === ecClass.schemaItemType) {
+      for (const m of Array.from(ecClass.getMixinsSync()).reverse()) {
+        baseClasses.push(m);
+      }
+    }
+
+    const baseClass = ecClass.getBaseClassSync();
+    if (baseClass)
+      baseClasses.push(baseClass);
+  };
+
+  while (baseClasses.length > 0) {
+    const baseClass = baseClasses.pop() as AnyClass;
+    addBaseClasses(baseClass);
+    if (baseClass !== this)
+      yield baseClass;
     }
   }
 }