+++ source/Metadata/Schema.ts
@@ -441,9 +441,9 @@ export default class Schema implements CustomAttributeContainerProps {
    */
   public schemaFromJson(jsonObj: any) {
     if (Schema.ec32 && SCHEMAURL3_2 !== jsonObj.$schema)
-      throw new ECObjectsError(ECObjectsStatus.MissingSchemaUrl);
+      throw new ECObjectsError(ECObjectsStatus.MissingSchemaUrl, `Schema namespace '${jsonObj.$schema}' is not supported.`);
     else if (!Schema.ec32 && SCHEMAURL3_1 !== jsonObj.$schema)
-      throw new ECObjectsError(ECObjectsStatus.MissingSchemaUrl);
+      throw new ECObjectsError(ECObjectsStatus.MissingSchemaUrl, `Schema namespace '${jsonObj.$schema}' is not supported.`);
 
     if (!this._schemaKey) {
       if (undefined === jsonObj.name)